requirements/base.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-ninja==1.5.3
djangorestframework==3.16.1
email-validator==2.2.0
pydantic==2.12.5
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
